pyproject.toml
[project]
name = "project-course"
version = "0.1.0"
requires-python = "==3.10.12"
dependencies = [
    "fastapi>=0.115.0",
    "numpy>=2.2.5",
    "opencv-python>=4.13.0.92",
    "orjson>=3.10.18",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "uvicorn[standard]>=0.34.2",
]

[project.scripts]
project-course = "project_course.main:main"
project-course-camera = "project_course.camera.cli:main"
project-course-api = "project_course.api.__main__:main"
project-course-api-dev = "project_course.api.__main__:dev_main"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "ruff>=0.11.7",
]

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
    "unit: unit tests for isolated modules",
    "integration: integration tests for service/data flow",
    "contract: API contract compatibility tests",
]

[tool.ruff]
extend-exclude = ["laboratory/legacy/motion-amplifier"]
line-length = 88
src = ["src", "tests"]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
